consolidate all modals
