+++ packages/app-shell/src/components/Modal/modals/Confirmation/getCopy.js
