+++ packages/app-shell/src/components/Modal/modals/PaymentMethod/index.jsx
@@ -1,12 +1,12 @@
 import React, { useContext, useEffect } from 'react';
 
-import { MODALS } from '../../common/hooks/useModal';
-import { UserContext } from '../../common/context/User';
-import { ModalContext } from '../../common/context/Modal';
+import { MODALS } from '../../../../common/hooks/useModal';
+import { UserContext } from '../../../../common/context/User';
+import { ModalContext } from '../../../../common/context/Modal';
 
 import StripeProvider from './components/StripeProvider';
 import Form from './components/Form';
-import { useTrackPageViewed } from '../../common/hooks/useSegmentTracking';
+import { useTrackPageViewed } from '../../../../common/hooks/useSegmentTracking';
 
 const PaymentMethod = () => {
   const currentUser = useContext(UserContext);
