+++ packages/app-shell/src/components/Modal/modals/PlanSelector/hooks/useHeaderLabel.js
@@ -1,4 +1,4 @@
-import { freePlan } from '../../../common/mocks/freePlan';
+import { freePlan } from '../../../../../common/mocks/freePlan';
 
 const useHeaderLabel = (isActiveTrial, planOptions, isFreePlan) => {
   let headerLabel;
