+++ packages/app-shell/src/components/Modal/modals/StartTrial/index.jsx
@@ -5,10 +5,10 @@ import Text from '@bufferapp/ui/Text';
 import Button from '@bufferapp/ui/Button';
 
 import { Error } from '../PaymentMethod/style';
-import { MODALS } from '../../common/hooks/useModal';
-import { UserContext } from '../../common/context/User';
-import { ModalContext } from '../../common/context/Modal';
-import useStartTrial from '../../common/hooks/useStartTrial';
+import { MODALS } from '../../../../common/hooks/useModal';
+import { UserContext } from '../../../../common/context/User';
+import { ModalContext } from '../../../../common/context/Modal';
+import useStartTrial from '../../../../common/hooks/useStartTrial';
 
 import { Holder, Content, Ctas } from './style';
